requirements.txt
# News Aggregator Dependencies

# RSS parsing
feedparser>=6.0.10

# HTTP requests
httpx>=0.25.0
aiohttp>=3.9.0

# AI curation
anthropic>=0.39.0

# Language detection
langdetect>=1.0.9

# Text processing
beautifulsoup4>=4.12.0
html2text>=2024.2.26

# Templating
Jinja2>=3.1.2

# Configuration
PyYAML>=6.0.1

# TTS
piper-tts>=1.2.0

# TTS server
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
python-dateutil>=2.8.2
tqdm>=4.66.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Text similarity for dedup
scikit-learn>=1.4.0
